# Cache unit conversion (13× speedup in halo catalogue access)