When converting units from a halo catalogue, a *lot* of time is spent doing the `dimensional_project` operation. This PR adds support for a caching mechanism to speed things up.

| Action | Time before (s) | Time with 783033a (s) | Time with 3f4f171   (s) |
|---|--|--|--|
| `f = pynbody.load(…)`| 2 | 2 | 2 |
| `halos = f.halos(); halos.physical_units()` | 0.5 | 0.5 | 0.5 |
| `halos[1]`| 25 | 25 | 25 |
| `[h.properties for h in halos]` | ~2000 | ~300 | ~150 |